+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * Bugfix: AHN_ENV and RAILS_ENV now do not interfere with each other when both are set, and ahn will boot in the RAILS_ENV if AHN_ENV is not set
   * Feature: The console can take control of a call
   * Bugfix: CallController#dial now blocks until all outbound calls complete
+  * Bugfix: Call commands timing out now raise a timeout exception in the caller, but do not crash the actor
 
 # 2.0.0.alpha2 - 2012-01-30
   * Change: Plugins no longer load dialplan/event/rpc/console methods using corresponding class methods